Solution to Calculate Employee Bonus

+++ Calculate Employee Bonus@@67877d61dc107486ad5c274f/code.c
@@ -16,7 +16,7 @@ int main() {
     }
 
     for (int i = 0; i < len; i++) {
-        float bonus;
+        int bonus;
         if (employee[i].salary > 50000) {
             bonus= (5/100)*employee[i].salary;
         } else {
@@ -24,7 +24,7 @@ int main() {
         }
         
 
-        printf("Employee ID: %d, Name: %s, Bonus: %.2f\n", employee[i].id, employee[i].name, bonus);
+        printf("Employee ID: %d, Name: %s, Bonus: %.2f\n", employee[i].id, employee[i].name,(float) bonus);
     }
 
     return 0;
